fix multiple bindings in a let
as it turns out, the way we were using it should be in the let* form
instead. the informal description of let's semantics in the Racket docs
suggest that, given an `[x e]`, `x` is not bound in any of the other `e`s
(other than the body). this would be annoying to implement without
introducing an explicit AST node for such a purpose.